experimental changes to js
WARNING: THIS IS BROKEN CODE. Don't use it on a live install.

I am trying to make a field that counts down characters live on the fly. For now, I just want to start with 200 and count down, and once I get it working I'll use AJAX to dynamically populate the "left" var.

But I can't even get the jQuery to work, and start counting down. 

Is this even possible?